+++ gratatouille-gradle-plugin/src/main/kotlin/gratatouille/gradle/internal/DefaultGratatouilleTasksExtension.kt
@@ -1,0 +1,19 @@
+package gratatouille.gradle.internal
+
+import gratatouille.gradle.GratatouilleTasksExtension
+import gratatouille.wiring.GExtension
+import org.gradle.api.Project
+
+@Suppress("unused")
+@GExtension("com.gradleup.gratatouille.tasks", "gratatouille", GratatouilleTasksExtension::class)
+abstract class DefaultGratatouilleTasksExtension(project: Project): GratatouilleTasksExtension {
+  init {
+    addDependencies = true
+    val defaultCoordinates = project.provider { "${project.group}:${project.name}:${project.version}" }
+    coordinates.convention(defaultCoordinates)
+    configurationName.convention(defaultCoordinates.map { "gratatouille${it.sanitize()}" })
+
+    project.configureCodeGeneration(this)
+    project.configureDefaultVersionsResolutionStrategy()
+  }
+}