+++ test-plugin/gradle-plugin/src/main/kotlin/testplugin/extension.kt
@@ -1,6 +1,6 @@
 package testplugin
 
-import gratatouille.wiring.GExtension
+import gratatouille.GExtension
 import org.gradle.api.Project